refactor: split server page exports + remove `pureCommentPlugin` (#476)
