+++ app/root.tsx
@@ -27,31 +27,17 @@ import {
   encodeFlashMessage,
   useFlashMessageHandler,
 } from "./utils/flash-message";
-import { RootLoaderData, useRootLoaderData } from "./utils/loader-utils";
-import { makeLoader } from "./utils/loader-utils.server";
+import { useRootLoaderData } from "./utils/loader-utils";
 import { cls } from "./utils/misc";
 import { navigateRefresh } from "./utils/misc-client";
 import type { PageHandle } from "./utils/page-handle";
 import { QueryClientWrapper } from "./utils/react-query-utils";
 import { ToastWrapper } from "./utils/toast-utils";
 
-//
-// loader
-//
-
-export const loader = makeLoader(async ({ ctx }) => {
-  const loaderData: RootLoaderData = {
-    currentUser: await ctx.currentUser(),
-  };
-  return loaderData;
-});
+export { loader } from "./root.server";
 
 // no need to revalidate `currentUser` since app refreshes on user session change (signin/signout)
 export const shouldRevalidate: ShouldRevalidateFunction = () => false;
-
-//
-// component
-//
 
 export default function DefaultComponent() {
   // hydration error for theme class (dark, light)
