+++ app/routes/decks/new.tsx
@@ -4,26 +4,14 @@ import { useForm } from "react-hook-form";
 import toast from "react-hot-toast";
 import { $R } from "../../misc/routes";
 import { trpc } from "../../trpc/client";
-import { makeLoader } from "../../utils/loader-utils.server";
 import { cls } from "../../utils/misc";
 import type { PageHandle } from "../../utils/page-handle";
+
+export { loader } from "./new.server";
 
 export const handle: PageHandle = {
   navBarTitle: () => "New Deck",
 };
-
-//
-// loader
-//
-
-export const loader = makeLoader(async ({ ctx }) => {
-  await ctx.requireUser();
-  return null;
-});
-
-//
-// component
-//
 
 export default function DefaultComponent() {
   const navigate = useNavigate();
