+++ app/routes/users/register.tsx
@@ -1,43 +1,20 @@
 import { Temporal } from "@js-temporal/polyfill";
 import { Link } from "@remix-run/react";
-import { redirect } from "@remix-run/server-runtime";
 import { useMutation } from "@tanstack/react-query";
 import { useForm } from "react-hook-form";
 import { $R, R } from "../../misc/routes";
 import { trpcClient } from "../../trpc/client-internal.client";
 import { encodeFlashMessage } from "../../utils/flash-message";
-import { makeLoader } from "../../utils/loader-utils.server";
 import { cls } from "../../utils/misc";
 import { navigateRefresh } from "../../utils/misc-client";
 import type { PageHandle } from "../../utils/page-handle";
 import { useTurnstile } from "../../utils/turnstile-utils";
 
+export { loader } from "./register.server";
+
 export const handle: PageHandle = {
   navBarTitle: () => "Register",
 };
-
-//
-// loader
-//
-
-export const loader = makeLoader(async ({ ctx }) => {
-  const user = await ctx.currentUser();
-  if (user) {
-    return redirect(
-      R["/users/me"] +
-        "?" +
-        encodeFlashMessage({
-          content: `Already signed in as '${user.username}'`,
-          variant: "error",
-        })
-    );
-  }
-  return null;
-});
-
-//
-// component
-//
 
 export default function DefaultComponent() {
   type FormState = {
