+++ remix.config.js
@@ -21,45 +21,3 @@ module.exports = {
       { ignoredFilePatterns: ["**/*.test.*"] }
     ),
 };
-
-//
-// see patches/@remix-run__dev
-//
-globalThis.__esbuildPluginsCommon = [];
-globalThis.__esbuildPluginsBrowser = [pureCommentPlugin()];
-globalThis.__esbuildPluginsServer = [];
-
-// tree-shake `export const loader = makeLoader(...)` from client bundle
-function pureCommentPlugin() {
-  const fs = require("node:fs");
-  const path = require("node:path");
-
-  return {
-    name: pureCommentPlugin.name,
-    setup(build) {
-      build.onLoad(
-        {
-          namespace: "file",
-          filter: new RegExp(
-            `${__dirname}/app/`.replace(/[.*+?^${}()|[\]\\]/g, "\\$&") +
-              ".*.tsx"
-          ),
-        },
-        async (args) => {
-          let contents = await fs.promises.readFile(args.path, "utf8");
-          if (contents.includes("makeLoader(")) {
-            contents = contents.replace(
-              "makeLoader(",
-              "/* @__PURE__ */ makeLoader("
-            );
-            return {
-              contents,
-              resolveDir: path.dirname(args.path),
-              loader: "tsx",
-            };
-          }
-        }
-      );
-    },
-  };
-}
